requirements.txt
# Mean Reversion Trading Framework - Python Dependencies

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# AWS
boto3==1.34.14
botocore==1.34.14

# Environment & Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Backtesting
backtrader==1.9.78.123

# Parallel Processing
joblib==1.3.2

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# Statistics & Analysis
scipy==1.11.4
statsmodels==0.14.1

# Progress Bars
tqdm==4.66.1

# Logging & Monitoring
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Data Validation
pydantic==2.5.3

# HTTP Requests (for data fetching)
requests==2.31.0

# Jupyter (optional - for analysis notebooks)
jupyter==1.0.0
ipykernel==6.27.1

# Type Hints
types-PyYAML==6.0.12.12
types-requests==2.31.0.10

# Linting & Formatting (development)
black==23.12.1
flake8==7.0.0
mypy==1.7.1
